+++ fivetran/framework/provider.go
@@ -17,7 +17,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
-const Version = "1.9.18" // Current provider version
+const Version = "1.9.20" // Current provider version
 
 type fivetranProvider struct {
 	mockClient httputils.HttpClient
@@ -103,6 +103,7 @@ func (p *fivetranProvider) Resources(ctx context.Context) []func() resource.Reso
 		resources.Webhook,
 		resources.Connector,
 		resources.Connection,
+		resources.ConnectionConfig,
 		resources.ConnectorSchema,
 		resources.ConnectorSchedule,
 		resources.Destination,